Make lqtnotifytalk true by default for users (second attempt)
This is the preference "Email me on replies to a thread I am watching".
Current stuatus is greatly counter-intuitive and breaks translators signup workflow
because translators expect an email notification of some sort after being approved.
<https://translatewiki.net/wiki/Thread:Translating_talk:Etherpad_lite/2_french_translations_to_modify/reply_(2)>

Don't enable in the middle of the night and remember to notify users
so that we can test and quickly revert if there are more bugs.

Change-Id: I5f21ea74ab60c4d4bff1061580a9b81eaf4a684f