Update our wikiupdate scripts
Replaced our custom parallelism with xargs -P which N updates in
parallel instead of running N commands at once and waiting for them
to finish before running the next batch. Lowered the number of
parallel updates from 10 to 5 since it is faster now.

As a bonus the code is shorter and output is quieter and clearly
highlights when repos are not updated.

Also made it to call subscripts from the same dir to avoid failures
if PATH is not set up correctly and so that correct scripts are run
when I developed this patchset in my own checkout.

Change-Id: Id7b606e21998387d04e9dabb7ff6edc0fc00d168