Disable oojs-ui and ve modules of VisualEditor extension
These have been moved elsewhere. Need a solution for adding these repos in
this file, or they have to be added as products, which isn't going to be
great for the maintenance overhead.

Alternatively, we could create a new config file that will contain only
JSON based groups with a repo config or something...

Change-Id: I6f58a0f3cd2161ea8ff9db8a826c6ddfd03c007e