Update patch set 4
Patch Set 4:

Raimond:

Yes, it's irritating that the daily updates of translations aren't yet working for JSON-sourced files, so we have to wait for the weekly (or more frequently) scaps to build them from scratch. This will be solved when Localisation Update is fixed to cope with this.

Note that this has been the case since VisualEditor switched over to the new i18n system in December, and this patch changes nothing in that respect.

Patch-set: 4