Update patch set 1
Patch Set 1:

If it quacks like a duck... it still might be only a rubber duck.

In case you are wondering, there clearly is a lack of flexibility in our repo&group configuration. Instead of living with it, or dealing with it, you suggest to force non-extensions to extensions by ignoring the abstraction layers present.

This is NOT a -1.

Patch-set: 1