Update patch set 1
Patch Set 1:

Not sure what you're suggesting, Niklas. I've marked this as -1 with a, what I think is a workable solution now, that I will implement in hopefully a few hours. I really don't want to add more products in the way that is done by this patch and it's sister.

Patch-set: 1