Update patch set 3
Patch Set 3: Code-Review-2

Previous patch set was used as proof of concept in https://gerrit.wikimedia.org/r/#/c/154798/. That worked for the conversion of the files.

Now blocking merge until the Pywikibot devs give me the go-ahead. Once this is merged, some work needs to be done on translatewiki.net immediately to get the converted files in the system. Rough steps:

- repo update pywikibot
- xsc
- b php processMessageGroupChanges.php --group=out-py*
- Process on Special:ManageMessageGroups
- cd /resources/siebrand
- repoupdate pywikibot
- bclean
- update /home/betawiki/config/bin/bxa to also export english
- bxa out-py*
- cp -r ~/export/pywikibot/* /resources/siebrand/pywikibot
- repocommit pywikibot
- merged the L10n-bot patch set.

Patch-set: 3