Update patch set 1
Patch Set 1:

It has so many translations only because it was in the 2011 most used list as well, it's no indication of importance. In theory it would be nice to split but then we need a way to show instructions above the message itself (which would be useful in many cases, granted).

Anyway, if you think removing it would make things worse then I'll abandon.

Patch-set: 1
Reviewer: Nemo bis <105@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0