Remove MoodBar from Extensions Used by Wikimedia
MoodBar is only used in a very small number of wikis,
its development is officially stopped and there are no plans to deploy it
to more projects. Removing it from the "used by Wikimedia" group
will reduce a bit of the load on the translators who want
to focus on translation of Wikimedia projects' software.

Change-Id: Idec321392f938fcb81c11a190c9360e75b44e594