Update patch set 3
Patch Set 3:

I've added some spaces in the name. I thought the optional line should be kept in sync with the intuition repo? When I added it as optional in the qqq.json my idea was that the message is not displayed in normal use of the tool, so I thought translators don't have to spend time on translating it.

Patch-set: 3
Label: Verified=0