Update patch set 3
Patch Set 3:

Our "definition" of optional is a message that does not contain such linguistic content that would need translating in most languages. It is mainly to allow languages to reach 100% translated without having to copy identical content for optional messages, which just makes it harder to change those later on.

Patch-set: 3
Label: Verified=0