Update patch set 2
Restored

This is the a only variant of Baluchi language is written in Latin.

The area this language is spoken in northwestern Pakistan and a Part of India, where their script is Latin.

I didn't find any text in language written in Arabic.

But it also needs a prefix or appendix since it's "Eastern Baluchi" and not "Baluchi".


We got a lot of requests to support Latin script of Baluchi:

https://translatewiki.net/wiki/Thread:Support/Baloc%C3%AD_Language_in_Latin_Script

https://translatewiki.net/wiki/Thread:User_talk:Nemo_bis/Baluchi_macrolanguage_%28bal%29_at_MediaWiki_core

But bal is separated into individual codes and bgp is the only one is written in Latin.

The script is like Brahui language:

https://en.wikipedia.org/wiki/Brahui_language

https://incubator.wikimedia.org/wiki/Wp/brh/Bunt%C3%A1k

Patch-set: 2
Status: new