Update patch set 1
Patch Set 1:

ISO 639-3 is an international standard for language codes. In defining some of its language codes, some are classified as macrolanguages, which include other individual languages in the standard. The category covers borderline cases where two language varieties may be considered strongly divergent dialects or the same language or very closely related languages (dialect continuums). It also is used when there are language varieties that are sometimes considered to be the same and sometimes different languages for ethnic or political rather than linguistic reasons.

We have generally seen that using macro-languages as locale names bring more bad (mostly internal unreset between speakers of the different languages) than good, and thus have a policy to not allow any new macro-langauges to be introduced in translatewiki.net, and subsequently MediaWiki.

Patch-set: 1