Update patch set 2
Patch Set 2:

Sorry for not replying to the last comment, which I had not understood. Are you implying that this wouldn't work on translatewiki.net, despite working in tests? I see the two instances use different language codes, is one of the two supported? What should be supported and where?

Patch-set: 2