Define Monobook, Modern and Cologne Blue as used by Wikimedia
The Vector skin is already defined in Used by Wikimedia - Main,
and it makes sense.

Monobook is the previous default and it's still
used by many active users, so it should be in Main.

Cologne Blue and Modern are installed, but less used,
so they should be in Extension Used by Wikimedia.

Change-Id: I3cb66cce8a701357d6ec3b0b64b72f2ed3d37d66