Update patch set 1
Patch Set 1:

im not sure about the Wikisource app (cant find it in the app store; googling only shows a non WMF app), but the Wiktionary app appears to still be in use, and usable on the latest release of Android), albeit the app description says it is not supported.  Is there harm in allowing more translations?  I cant see many translatable strings in the app.  And the app may not be any better than the desktop skin, in which the product should be pulled of the app store ;-)

Patch-set: 1
Reviewer: John Vandenberg <1718@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0