Update patch set 1
Patch Set 1:

I will leave this open for a while (but not for long) in case anyone wants to reanimate the Wiktionary app and do releases. However, I am not going to poke people to do that.

Removal from translatewiki.net is not permanent action, it is possible to add it back later if appropriate conditions are met.

Patch-set: 1