Reduce error output from repocommit
The previous one supressed failues from git commit with "|| :".
This made it so that get review is executed even if a commit is
made and that then fails with error (git push doesn't).

Fixed by using if-then so that push/review is only executed if
a commit is created, and that failures from push/review will be
detected.

Change-Id: Icd8fabe5a4397f638e2514fce46a9bb78c215823