Split API from Echo
There are many API messages in Echo, so to make it easier
for translators, they are now in a separate group.
This is done similarly in Translate, ContentTranslation, and Flow.

Must be merged along with the change in the Echo repo at
I1d3baea708107a7f15bf19671f7386eaf7e33a1b

Change-Id: Iecedfe4cb9dc8e62a446a3e1c415a79e116ca27e
Depends-On: I1d3baea708107a7f15bf19671f7386eaf7e33a1b