Update patch set 3
Patch Set 3:

Thanks :] I submitted another stab at it. I don't know how to test any TWN changes outside of production so this might not work or you might prefer not to accept the changes. Let me know if this is the case and I'll just drop the auto-merge part of the patch

Patch-set: 3
Label: Verified=0