Update patch set 2
Patch Set 2: Code-Review-1

So, this project did not have any export threshold. In the new system the implicit default is 35%. We could take this as normalization so that it is consistent with other projects, or we could set it to 1% to act almost as the old behavior, or we could fix repong to allow not setting export threshold. What would you prefer?

Patch-set: 2
Label: Code-Review=-1