Remove MediaWiki extension BoilerPlate
This is supposed to be a copy-paste-able boilerplate template for
creating your own extension. As such, it doesn't make sense to have
i18n files other than en and qqq in the repository -- the first
thing a developer has to do when basing an extension on BoilerPlate
is to delete them.

Localisable information about this extension should be kept on
MediaWiki.org, not in the repository.

Bug: T156569
Change-Id: I1f688f4f60507509e66f1c5240d4bc0b48fb3f2d