Lint puppet material using rake
The CI tests run for the repository are hardcoded in Wikimedia Jenkins
job, that makes it hard to reproduce locally and difficult to adjust.
This change propose to define the commands being run directly in the
repository and use ruby rake to run them.

Currently the CI runs:

translatewiki-puppet-validate: runs 'puppet parser validate' against
each of the .pp files found in the repo.

translatewiki-puppetlint-strict: runs puppet-lint with custom log format
and making sure it fails on warnings.

The first can be replaced by the ruby gem 'puppet-syntax' which does the
puppet parser validate but also check erb templates and hiera yaml
files.
puppetlabs_spec_helper provides the rake tasks for us.

puppet-lint ship with a rake task as well.

List the dependencies we need in Gemfile. Allow to easily change puppet
version with the env variable 'PUPPET_GEM_VERSION' would let us have a
job for puppet 3.7.x and another for puppet 4.8.x.

safe_yaml is added as an explicit dependency, else puppet 3.7 fails to
load when running with ruby 2.2+

The installation is made using bundler, a package manager for ruby.
Instructions are in the Rakefile but in short:

 apt-get install bundler
 bundle install
 bundle exec rake test

Running 'bundle exec rake' shows up all available tasks.

Bug: T158544
Change-Id: I6789dcf2e8cf90d1e535125989ecdb1ea9756907