lint bash script from composer test command
I would like to drop the Jenkins job translatewiki-shelllint which has
the hardcoded command:

    find bin -maxdepth 1 -type f -print0 | xargs -t -n1 -0 bash -n

That does not offer much liberty to developer of translatewiki.git since
any tweak requires a change in the Jenkins job which is not
straightforward.

Port the functionalitly to php as bin/lint-bash
It reads all files in the same directory and pass each to 'bash -n'. If
any fail, it exits 2.

Add it as a build step in 'composer test'
Pass the new script explicitly to parallel-lint which expects files to
be suffixed with '.php'.
Same for PHP_CodeSniffer.

Usage:

 $ composer test
 ...
 > bin/lint-bash
 All bin scripts pass `bash -n`.

Bug: T160394
Change-Id: Idd3b7f0a05d09dc9df58063fc4fa6438180f05eb