Update patch set 1
Patch Set 1:

Really sorry for messing with this patch. My fault :-(

But I still see no reason to ignore the message. The content is visible in the MediaWiki namespace and a translated first sentence is helpful for every sysop who is not fluent in English language.

Furthermore we have some more message where you could argue the same way to ignore them.

Patch-set: 1