Remove project Europeana
Per Ad Pollé, Collections Manager Europeana, Senior UGC Projects Coordinator:
"This is to let you know that the Europeana 1914-1918 platform will be migrated
to the Europeana Collections platform. This also means that we will not make use
any longer of translatewiki.net for our translation procedure. We will therefore
proceed by undoing all connections to this activity and site.

We would like to thank you and all translators for the help and work carried out
in this process which helped to make our project more successful."

Change-Id: Ifffc1f12f8f91084b056e537a114d9fb1cd386ce