Update patch set 1
Patch Set 1:

It has been suggested that instead they could require that the language meets the export threshold and gets exported for core.

Or if you continue to require this group to be complete, then there is a bit more work, but at least they are not getting stuck on not meeting the export threshold.

Patch-set: 1