Update patch set 1
Patch Set 1:

The id is a bit on the long side but acceptable. We have ia-upload where ia- is part of the name rather than a prefix, but we could use it as a sort of shared prefix anyway.

The CamelCase name is consistent with the Phabricator project.

Patch-set: 1
Reviewer: Nemo bis <105@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0