Enable the new recent changes filters
This requires disabling the "clean" listing from clean changes. In
practice this is not an issue due to use script we have that makes
the user tool links compact with JavaScript. I fixed that to use
the wikipage.content hook so it works well with the new filters.

This also removes the annoying bug that headings are duplicated
in the recent changes listings.

Change-Id: I9f14e03320b6f478f19a2a0c0a1bbd2d980be773