Add rudimentary state synchronization
The purpose is to prevent accidental overriding of commits
that happen between our last synchronization to translatewiki.net
and us exporting those changes.

For example:
1. Update read-only repo
2. Synchronize into the wiki
3. Update read-write repo
4. Export and commit translations

In step 3 the read-write repo could get ahead of the read-only
repo. Now, if possible, we read the git commit from step 1 to
reduce the likelyhood. This is not completely safe yet, because
there is no guarantee that step 2 has happened. For humans this
should be fairly obvious, because our scripts always alert after
step 1 (which is automatized) that there are pending changes
to check. For fully automatic exports this is not safe enough.

Step 4 is amended to attempt automatic rebase before pushing to
avoid failing on non-conflicting changes.

Removed update-reset-repo which is now enhanced and inlined in
clupdate-git-repo.

State synchronization is not really possible when we are committing
to a different branch or repository. Different branch is detected
automatically and state is not synchronized then. For different
repos there is new option no-state-sync that can be set to disable
synching.

This commit uses the null coalescing operator (??) available in
PHP7 and HHVM.

Bug: T48833
Change-Id: I979cbc20f31cf827a58bfd0fdd9ca2a7803ae345