Update patch set 1
Patch Set 1:

I realize this will spam all of your inboxes a lot, so feel free to remove yourself from this patch.

My understanding is that I am only making the implied license explicit [1], so there is no need to have an approval from everyone. I am giving the opportunity to object if someone was under a different assumption and are not okay with the license, in which case we will look into how to solve that. If you think this approach is flawed, let me know.

[1] Everything hosted in Gerrit must be under an open source license, and GPL-2.0-later is the standard Wikimedia software license. This license is also mentioned in many, but not all, of the files.

Patch-set: 1
Reviewer: Nikerabbit <93@e9e9afe9-4712-486d-8885-f54b72dd1951>