[Cite] Update optional and ignore messages
Most of them have no real content for translation -> ignore

Translation identical to English for the following messages are deleted on translatewiki now.
https://translatewiki.net/wiki/Special:Translations?message=cite_references_link_many_sep&namespace=8
https://translatewiki.net/wiki/Special:Translations?message=cite_references_link_many_and&namespace=8
https://translatewiki.net/wiki/Special:Translations?message=cite_references_link_many_format_backlink_labels&namespace=8

For some languages valid translations are useful and necessary.
Therefore it is no option to ignore them,
In any case if someone will again translate them identical to English it
is not useful but it does not hurt on a technical base,

Change-Id: I622f398a213582f821e9fde22281ec9a1a43e818