Simplify descriptions of some extensions' aggregate groups
As part of T189982, some MediaWiki extensions were split
to a main group and an API group. Several more extensions
will probably be split this way.

Nikerabbit suggested to simplify their configuration and
to reuse their own description for the description of
the aggregate group instead of defining repetitive
descriptions on the [[Group descriptions]] page.

This patch defines some of the aggregate groups this way.
Some of the groups, which have a more meaningful description
on the [[Group descriptions]] page are left with it,
such as VisualEditor and UniversalLanguageSelector.

Bug: T189982
Change-Id: I65f1a7ff737ce8b45ad35151c1be8daf9b360c15