Ignore some messages keys to be added in Ie2483e12
Those keys will be only used on meta.wikimedia, probably test.wikipedia
and a couple of beta cluster wikis. No need for translators to invest
time translating a string that isn't showing on non-English wikis,
although adding the keys is still a good idea to unbreak log entries and
summaries.

Depends-On: Ie2483e1213f043269aaa5c49815b9d6370b50575
Change-Id: Icabb80d063743ff72475bb9b599e9ca9ccc8adf7