Update patch set 1
Patch Set 1:

> (1 comment)

"vls" does not seem to be used as a language code in Django, so I assumed it also had to be translated to something it can understand. Although "nl_BE" does not appear in the localization files either, it respects the expected format, and language fallback to nl should work as expected.

But we currently don't have translations in this language anyway, so happy to remove this if you think it is better.

Patch-set: 1
Reviewer: Pintoch <5361@e9e9afe9-4712-486d-8885-f54b72dd1951>