Update patch set 1
Patch Set 1: Code-Review-1

> Are you sure that works as expected? MessagesNdq was added last
 > week but I guess an entry in Names.php is necessary. But currently
 > no time to test my guessing.

You are a bit too quick :)

I was about to post the following comment both here and at the corresponding core patch:

I'm jumping ahead of time a bit. The translation of the most used messages to this language is progressing quickly, and even though the first export of messages from translatewiki to core hasn't been done yet, I am making the patch to be ready for the time that it happens. The core Names.php patch should be merged after the export, and the translatewiki patch should be merged after that.

Patch-set: 1
Reviewer: Amire80 <54@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1