repong: Implement 'push-branch' option
This will be used by the repong/CommitCommand step when called
from localisation export script.

This is different from the 'branch|export' option because 'push-branch'
will only be used for the 'git push' command.

The 'branch' option is also used for the clone, checkout, and rebase
steps. The problem with that is that it means the 'translatewiki'
pull-request branch is never up to date and piles up old commits.

Change-Id: I00d3eee409263815d76c136c55dad2023c0d2b20