Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I suggest to rename the label of either this or the old tool. Otherwise they cannot be distinguished in the group selector.

Good point. And that makes me realize that this might be more complicated than I thought, because this tool is *replacing* the old one. We could conceivably usurp all of its messages on Translatewiki.net — if that's a thing that can be done?

Patch-set: 3
Reviewer: Samwilson <828@e9e9afe9-4712-486d-8885-f54b72dd1951>