Add common known hosts file to be used for exports and imports
Renamed persistent-ssh to git-ssh-wrapper.

Updated the git-ssh-wrapper file to add a known_hosts file to the
ssh commant and disable IP level checks. The known_host file has
been added under the data folder. Also disabled hashing of known
hosts to allow the known_hosts file to be reviewed.

Moved the GIT_SSH command into repocommit so that it is used for
all exports / imports.

Bug: T226739
Change-Id: I4494218b183ccda46bc627dca4f54b91271a9ec7