Move apihelp messages in ConfirmEdit to separate files
This is one of the last extensions to undergo this conversion,
and it's a bit different and perhaps requires a more careful review.

ReCaptchaNoCaptcha is very small and not used by Wikimedia,
but it's converted anyway for consistency.

The corresponding extension patch is
I66faae6fd4ff447327587c89ad2a1704edd1b356

Change-Id: I1716bca6d46b25684b9804b14fc850f3e9141965