puppet: fix keyholder
Neither the service nor the key was being applied. Enable the
service by default, since it does not require any keys to function.

Add a small hack to site.pp that allows defining the keys in hiera
data. Alternative would be to add a new class to keyholder or in a
new module that does that.

Change-Id: Ia28bccf4afe90e7143a52cc398737412a0a5fc75