Replace "this is not production site" warning with canary warning
Since dev.translatewiki.net is no longer on the same server, the
old warning is no longer needer. Replace it with a canary warning
instead. Also add nice striped background when canary is in use to
make it really obvious. Color looks pretty and both white and black
text is still readable on it.

Bug: T247992
Depends-On: Ic65c59a05e90794ee740fe5e33148ed69d3d8ea7
Change-Id: I27a3656d5d546ce20230e579bbcf57abba9ff2f1