Move translatewiki.net newsletter preference from Translate
Simplified the code, no need to check for config when we know for
sure that emails is enabled. Also all new users must have a confirmed
email address, and doesn't harm if the others see the preference.

The message translate-pref-newsletter should be added to
https://translatewiki.net/wiki/MediaWiki:Betawiki-messages
once the message is no longer with Translate extension.

Bug: T134633
Change-Id: Idb9cb611c60b69836d1ceeece294927a6ff10f01