[Timed Media Handler] Add more optional keys about audio formats
For each message of a format (like 2160.mp4) there is also a -desc-
message. Added the -desc- message also as optional, because there are
not so many words.

The description messages with "low bandwidth" or "Small" not optional.
That allows to translate at least these words.
Reorder the keys to have the normal and the desc message in one line for
easier look. I have not resorted the audio formats to keep the diff
simpler.
Also add some timedmedia-resolution-* messages

Change-Id: Ib70da5abcc7fb26cc624f56548af709b12676455