Move code ciw to ojb in LanguageSettings
PLEASE NOTE:
Merge and deploy this in a way that takes bug T281053
into account.

This is done because the current translations into ciw
are actually more appropriate for ojb.
For more information, see bug T280461 and the discussion at
https://translatewiki.net/wiki/Thread:User_talk:Aandeginini/Ojibwe_translations

If anyone wants to add specifically ciw in the future,
it's possible.

Bug: T281053
Bug: T280461
Change-Id: Ib8e6cc7b51334b659b4e19dc647b87254b299c36