Add Malay in Arabic script
Requested at
https://translatewiki.net/wiki/Thread:Support/_ms-arab_Activate_new_language:_Malay_(Arabic_script)_-_%27%D8%A8%D9%87%D8%A7%D8%B3_%D9%85%D9%84%D8%A7%D9%8A%D9%88%27

This is a valid variant, with somewhat limited, but
common enough usage. It can be particularly valid as a language
of user interface and values in Wikidata.

This also makes it possible to create MediaWiki sites
with this language as content, although for creating Wikimedia
sites in it, it will have to be discussed separately with
the Language committee.

Change-Id: I297483a5cd6cc0171f4cf7c328e9e63874685256