Fix the Fon language autonym
I added a generic *"Fon" autonym at a hackathon, so it was done
in a rush. I was working on it with a native speaker, who wasn't
entirely sure about the correct diacritics to use, so we wrote
a generic name.

I later found in the Höftmann & Ahohounkpanzon dictionary that
the correct spelling of the autonym is "fɔ̀ngbè". It already appears
as such in language-data, but we forgot to update it in translatewiki.

Change-Id: I1bdd15e666b405cb21f17c17e95a33b0a0093ad8