[Wikibase - Repo] Remove the "ignored" status from many messages
They were marked as ignored long ago because of T40675.
The explanation was that the core API doesn't allow localization,
but since then the core API started allowing this. Besides,
despite having "api" in the keys, these messages are actually shown
to end users and not just to developers. Example:
https://he.wikipedia.org/wiki/File:Error_while_attempting_to_link_a_redirect_page.png

When this is merged and deployed, I'll remove the "{{Ignored}}"
mark from their qqq.

Bug: T287346
Change-Id: I468bf8d7eb0cb188e8bf032d1148d5fbd67d4220