Add new project for Oppia Android translations
This is introducing a new translatewiki project for Oppia Android per
the instructions listed at:
https://translatewiki.net/wiki/Setup_of_a_new_project.

For reference, https://github.com/oppia/oppia-android is repository
corresponding to the project being added. Note that this is being added
as part of the current Oppia translation group, though we don't intend
to reuse translations across the two platforms.

The initial Oppia Android group will be restricted to just a few
languages since we don't plan to launch full support for all languages
Translatewiki supports in the initial app launch.

Change-Id: Ib8f83cebd9aff6f58cdb11b260c22c2a50dc2cba