puppet: relax nginx rate limits
We got reports that users are hitting these limits in normal use.

Increase the maximum number of php-fpm workers so that it is higher
than the burst value, so that one client cannot use all workers.

The average memory use of worker is about 100M, so 30 of them would
use 3G of memory, which is acceptable.

Change-Id: I9efde6d558d4f8c20a1fb5ce41afb50e60e79eb1