Commit support for backports
This includes the following changes for clupdate-git-repo:
* Always do repository state reset. Previously we would miss checking
  out the requested commit when cloning the repository.
* Use explicit tracking configuration (-t) which requires to explicitly
  name the local and tracking (upstream) branch. Use -B to always reset
  the branch. Previously, if anything had changed the tracking branch
  to something else, it would not have gotten reset by the script. This
  never happens when using these scripts, only if someone was manually
  poking things, but better make this more robust and future proof.
* Remove some use of subshells, which were there mostly to contain
  changes to current working directory, which is now explicitly used
  for the rest of the script after cloning part.

The motivation for this change is to ensure we setup the tracking
branch correctly, so we don't submit backports to the wrong branch.

In the commit command, we now rely on the tracking setup instead of
specifying the branch and upstream (unless using branch|export).

Backporting is explicitly disabled for repositories with push-branch,
pull-branch and branch|export configuration, since it is not possible
to determine in which branch to submit changes without further
refactoring.

Bug: T272830
Change-Id: I0bda75b024d0c76d9d58fdebd6392cad560ddbae