Rename language gur to "farefare"
The name "gurenɛ" is used only in Ghana and refers only
to one dialect. This language is also spoken in Burkina Faso,
and the name "farefare" covers all the dialects in both
countries. The activity in this language in translatewiki.net
and in the Wikimedia Incubator is higher recently, and this
was requested by the contributors to this language.

Confirmed with the book
De la phonologie à l’orthographe : Le ninkãrɛ au Burkina Faso,
Idda et Urs Niggli SIL, 2007
(page 6).

Change-Id: I917889228fa1554963adb1f3517d4be77bde580f