Toolhub: Exclude qqq from export
Exclude the qqq message set maintained on TWN to provide translators
with contextual information about translation units from export back to
the Toolhub repo. These documentation messages may not compile from po
to mo format due to mismatches between the data needed on TWN and the
source message requirements. This most commonly occurs when documenting
a message that is parameterized.

Bug: T296301
Change-Id: Id532aebad2545dd6fcd1c30916460af0c904f687