Fix the autonym of Okinawan (ryu)
The previous name was added in 2008, and it's
not actually the autonym, but its IPA pronunciation.
This language uses several orthographies based
on Japanese writing systems, and several linguists
whose expertise is Ryukyuan linguistics recommended
that I use this Hiragana-based name as the most
neutral and understandable. I'll also add aliases
to ULS to ensure that it's findable.

Change-Id: I9de58fb0d20aa6bf878cb1d4252475e1d6a972ec