Add Campidanese Sardinian (sro)
Requested at
https://translatewiki.net/wiki/Thread:Support/Request_for_Campidanese_(sro)_qualification

Also add fallbacks to Common Sardinian and Italian.

Autonym according to
Giovanni Melis,
"Fueddariu Sardu Campidanesu - Italianu",
Selargius, 2004

Comment: This enables the language for localization
on translatewiki, but doesn't necessarily imply
approval of creating the Wikipedia in it.
This will have to be discussed separately
by the Language committee.

Change-Id: I379ecbf60c99d50946568bf0291f80d144e33ead