Mark a few more Liberty skin messages as optional
Much like the core message "unit-pixel", these messages don't need usually to be translated and should really be translated only when the unit-pixel message is also translated, e.g. commonly when a language is written in a non-Latin script.

A notable exception is the "liberty-layout-select-1200" message, which does contain actual translatable content (the string "Default") and should always be translated when possible.

Change-Id: I82e0f66608467f1358dafad7e64175db900467c8