[Blockly] Remove a message from optional list
This key actually shouldn't be optional. It was
marked as optional by mistake and I subsequently
added it. I've gone over all thos optional messages,
and all of them are URLs that should be optional,
except this one, which is not a URL.

Follow-up to
I7302ba7f99808ac6b606fe96ecf18e696aa5d2cd

Change-Id: I417cdfbce9840279326286480a5e7e2103b00ad6