Add Igala (igl)
Requested at
https://translatewiki.net/wiki/Thread:Support/Igala_Language

For now, adding with the autonym "Igala", which appears
on Ethnologue. There are several variants with tones,
e.g. Igálá, Igáláà, Ígálâ, etc., but no apparent single
standard, so staritng with something simple that
is safe and easy to find.

Later, if a standard emerges, the autonym can be updated.

Change-Id: Id433b439c7d3e6a3e6a1f52b9fed5245317b4fe0