Split Manchu to Latin and Mongolian scripts
Currently, translations are done mostly in the Latin script,
and some are also done in the Mongolian script.
Both are used in real life, too, so both should be
available for localization.

This patch is the first step to splitting them. Since there
are just a few Mongolian-script translations, I'll move them
manually in translatewiki to separate them from
the Latin-script ones.

Once they are fully separated in translatewiki, all
the Latin-script translations will be moved to mnc-latn
using a bot.

Bug: T284043
Change-Id: I9065605fb4867cb39b09ea50b56c228740ae39f4