mw-ext: Ignore `grouppage-wmf-supportsafety`
Ignore message `grouppage-wmf-supportsafety` from the WikimediaMessages
extension. As it happens with the `wmf-officeit` group, this is a local
user group that exists only on Meta-Wiki, thus there is no need to
translate the group page key since it's not used anywhere else.

Documentation changes will be added to the respective qqq.json pages in
translatewiki.net, including the `{{ignored}}` and `{{notranslate}}`
notes, after approval.

Bug: T330514
Change-Id: I2bc5810ac16ca7384218097eaacb18b65531a184