[CampaignEvents] Move Rest messages to a separate file
This is similar to the kind of thing that was done
for many extensions in bug T189982, but here it's
done for Rest. Other than the title, it's largely the same.

Also move this extension from Upcoming to Advanced,
because it's already deployed in production
(but only on Meta, hence not Main).

Should be merged together with
Ib1b6861789a8ea04c77a4f150ce787dfd4988ac7

Change-Id: I40354f7006908fdc85f9dddd1a7c18d12ad69d48