Fix Gerrit auto-merge for intuition
Merged intuition-tools to wikimedia-tools, since the paths are mostly
the same. Maybe the github tools could be moved to wikimedia-misc-tools,
but that requires expanding int-* patterns for exports. Not done here
to keep the patch small.

I'd merge intuition to wikimedia-tools as well, but it has
export-threshold of 1 and needs checking if that can be removed.

Change-Id: I4549efd9eeee73f869d200a346fc3982531f6032