Add Karekare (kai)
There is an active Incubator in it, so it should also
be available for localization.

Add Bole and Karai-Karai as each other's assistant
languages because they are related, and Hausa
as a major language of the region.

Autonym according to the book
"Dindeno, asumka waɗo mabo Karai-karai"
Yobe Languages Research Project 2009
https://ucla.app.box.com/s/xvpgk2ecl73nwnjyyo7g80fsr96mtsbz

Change-Id: Ic94ca89f7856b930eefe93559e11e84640ccb263