[MediaWiki] Remove two "@deprecated" comments from optional messages
In I788ec328f75230768bffe28998d0b60e0427d4d9,
the messages "changed" and "created" were marked as
"@deprecated. Remove in MediaWiki 1.23."

We're in 2023, and they still seem to appear in the code,
so they probably aren't deprecated.

In addition, because of changes in this file, one of those
comments erroneously appears next to the message "content-model-css",
which is definitely not deprecated.

Change-Id: I23e79918477f41a69049576b4c5682078654c77c