[PageTriage] Remove many optional keys
Many keys were defined as optional because this extension
is aimed at the English Wikipedia. However, this is done
inconsistently: there are many other messages that aren't optional,
and all of these that are marked as optional now
have translatable content. Translating this whole extension
is low priority for all languages, but making the messages
optional is not so necessary; it's better to simply not
include it in the "Used by Wikimedia" group, which
is already done.

Only a few optional messages are left, because they actually should
be optional.

Change-Id: I378c73ba2dd456628c29e6f808e93f369581c1d0