Limit stats-newcorelangs to only actually new files
The page https: //translatewiki.net/wiki/Translating:MediaWiki/New_languages
is supposed to list new languages in core.
Currrently, what it actually does is listing new .json files
under languages/i18n/, including its subdirectories,
so it also shows new files under languages/i18n/exif/,
and that is not so useful.

The `git log` line, which generates the output was written
before the exif/ directory was created, so it makes sense
that it didn't take subdirectories into account.

This change only lists *.json files that are immediately
under languages/i18n and not in subdirectories.

Change-Id: I66f71112ffd9710da11adb75e4ca265528a904c2