Add tag cache to MediaWikiTopMessageGroup
The job queue sometimes get "stuck" causing message group sync failures.
After investigation, it has been determined that a common cause was a
MessageGroupStatsRebuildJob that takes hours to run. This only happens
when MediaWikiTopMessageGroup is the type of one of the included groups.

Profiling indicates that over half of the time is spend in getTags,
which is relatively slow when called for thousands of message groups
repeatedly. Since tags are not changing during one job, a simple cache
speeds up the execution considerably.

The changes here are intentionally minimal. There is another patch
(I78eb0fc015a3618f255295783355c7047c1250ed) to move this to Translate
extension where the code can be better maintained.

Bug: T208430
Change-Id: If39ae44c6d779fcd6dfd504698ec1034a36c0274