Remove Commons Wallpaper Gnome Extension
This is a project I made, but it is not working with the latest
versions of Gnome, and I don't have time to fix it/maintain it
at the current time. I will hopefully be able to make it work
again one day, but it is not in the foreseeable future.

Change-Id: Ic5bcae9de6ac5fa487f69ebd537f4d3f52e9f4ea