Update Northern Tujia autonym
The name that was used till now, "pi tsi kha",
is not the written name of the language, but
an IPA transcription of the name of the ethnic group.

The name "Bifzivsar" is the one that was requested at
https://meta.wikimedia.org/wiki/Requests_for_new_languages/Wikipedia_Northern_Tujia

It is also used in academic literature, for example
"Ethnicity as Policy Expedience: Clan Confucianism in
Ethnic Tujia-Miao Yongshun" by Chih-Yu Shih, published
by Taylor and Francis.

Change-Id: I8ef9d601fc72344ae82cfa4c220f27ff8c35b090