Add the "Extensions used by Wikidata" group
This is a cross-section of extension used on Wikidata.
Some of them were made specifically for Wikidata.
Some others, like Babel and ULS, are general,
but heavily used on Wikidata.

Wikidata is completely covered by the
"Extensions used by Wikimedia" groups, but
there were several requests to make a group
that targets specifically Wikidata, and it's
easy to make, so why not.

There are, of course, many more extensions
that are important for Wikidata, such as
DiscussionTools, Translate, Echo, etc.,
but it would probably be overkill to put
them all here. However, if lots of people
think that it's useful, it's easy to add
them later.

The description messages is not made yet,
but if this patch is merged, I'll quickly make it.

(This patch also adds spaces to the label of
"Wikibase Quality Constraints" for consistency with
other extensions.)

Bug: T372399
Change-Id: I62126534a63ced7ef24721531e7398f8328c2002